moved functions into seperate module and implemented pavel's commments
